crypto/openssl: handle case where there are no/NULL PKCS7 signers
Static analysis fix. sk_PKCS7_SIGNER_INFO_num can return -1 if sk if
pkcs7->signer_info is NULL, or 0 if there are no signers. Either case
will lead the following for-loop to immediately exit.

rc does not have an initial value. Therefore if this occurs, a comparison
will then be made against garbage data. This may lead to an incorrect
return of an error.

This patch fixes the potential issue by checking the number of signers
before entering the for-loop, and properly reporting the reason for
exiting early.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>