update: fix potential NULL pointer dereference with append updates
Static analysis fix. Append updates seem to handle timestamp checks in
a different way than regular updates. If auth_data->current_time is NULL,
it skips making a timestamp check, but later in the code unconditionally
makes a timestamp check to set the new time.

This patch conditionalizes the second timestamp check on if the timestamp
is NULL. If so, use update_time instead.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>