test/Makefile: only build tests as static
libstb-secvar is only really intended for inclusion into other components
rather than to be linked against as an actual library. Therefore, there is no
benefit to the tests linking against it as a dynamic library, and only adds
more complexity to the building and running of tests (namely, LD_LIBRARY_PATH).

Therefore, just build tests with libstb-secvar included statically, and
dynamically link against any other dependencies that should already be in the
ld path.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>