test/Makefile: move LDFLAGS to the end of the compile line
Apparently some compilers care about what order the flags are in. The -lcrypto
flag was not being picked up by gcc in the github actions (forecoming) test
container, which led to linker errors.

Reordering the flags appears to make it work for both my host compiler and
the actions compiler. I hope it does for other systems too.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>