log: add log global variable to control logging output
Logging is currently hardcoded to always output all messages. While most
consumers of this library should be including files directly and not via the
actual .so/.a output, it should be controllable by anything that links against
these files.

Namely, this allows the test cases to not output a lot of garbage and error
messages, even for passing tests.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>