test: rewrite tests to reduce output, consist use of assert()
Some of the test cases outputted way more information than is necessary.
In the case of CI, we should only see the information we need to debug a
problem. Furthermore, a test case binary should return a different return
for successes and failures -- currently some test cases always return 0.

This patch:
 - rewrites the logic to use assert() for all checks
 - removes all unnecessary printfs
 - disables prlog output from the library to limit excess output

This patch also include some helpers like assert_rc() to automate checking
return codes and print out a helpful message that translates error codes
to SV_ error strings.

The use of assert handles both the output of what failed, as well as returning
a nonzero return code, so that a CI or build system can detect a failed case
and handle that appropriately.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>