fix portability concerns with size_t printfs
Apparently size_t is a concern for printf, as %lu may not be the correct size
to printf as on all platforms. There are two options to address this:
 - cast size_t to unsigned long
 - use %zu

This commit uses the former option, as the second option is apparently also
nonstandard for compilers, and also is dependent on C99. Casting seems to be
the "most" portable option, other than not using size_t, printf, or probably C.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>