build: rework cppcheck targets, relax some checks
There's no need for cppcheck to be run separately for the library source and
the test cases, and in fact it causes an unusedFunction check to fail. By
including all the files in the cppcheck pass, it correctly detects that the
"unused function" is being tested in a test case.

Ignore certain cppchecks as they can not be (easily) addressed. This includes:
 - variableScope: in an effort to support as many C compilers/standards as
     possible, do not warn on variables that could be defined in an inner
     scope. Some compilers are still picky about declaring variables early.
 - missingInclude: the include paths in this project are messy at the moment.
     Suppress this one until time can be spent to fix this treewide.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>