esl: fix next_esl_from_buffer sanity check for current esl pointer
When iterating through esls in a buffer, the "current" input pointer must be
within the bounds set by the start of the buffer, and the supposed buffer
size.

The current check imposes an impossible failure state, as the current pointer
cannot be both *before* AND *after* the supposed buffer range of address. It
is likely this was meant to be an OR check, to see if it falls outside of one
of the ranges.

Change the inner range check from && to ||.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>