log: disable prlog for coverage testing objects
The current implementation of prlog introduces a branch for every call,
and only one of those branches will ever be taken during the lifetime of
a test case. This bloats the unreached branch count in coverage testing.

This patch adds a compile-time option to replace prlog with a no-op
macro, to completely remove it from the build.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>