external: move endian/endian.h inside a top-level ccan/ directory
To keep better track of the source of dependencies, endian.h should be
in a directory path that names the original source, as it is helpful
for namespacing headers.

This also matches that pattern that was used in secvarctl, so now
secvarctl can remove carrying endian.h itself and just use the version
provided by libstb-secvar.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>